Add improved bot logic
